Squash unreleased execution API version entries into 2026-06-16
